[Refactor] 폴더명 Group → GroupRoom 리네이밍 통일
