+++ src/main/kotlin/digdaserver/domain/group_room/domain/entity/GroupRoomRole.kt
@@ -1,4 +1,4 @@
-package digdaserver.domain.group.domain.entity
+package digdaserver.domain.group_room.domain.entity
 
 enum class GroupRoomRole {
