+++ src/main/kotlin/digdaserver/domain/todo/domain/entity/Todo.kt
@@ -1,6 +1,6 @@
 package digdaserver.domain.todo.domain.entity
 
-import digdaserver.domain.group.domain.entity.GroupRoom
+import digdaserver.domain.group_room.domain.entity.GroupRoom
 import digdaserver.domain.user.domain.entity.User
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
